+++ vyper/codegen/return_.py
@@ -1,6 +1,7 @@
 from typing import Any, Optional
 
-from vyper.codegen.abi import abi_encode, abi_type_of
+from vyper.codegen.abi_encoder import abi_encode
+from vyper.codegen.abi_types import abi_type_of
 from vyper.codegen.context import Context
 from vyper.codegen.core import (
     calculate_type_for_external_return,
